Create mars_lander-level_1.rb
Started trying to work out where the flat ground was from coordinates, but realised this puzzle was straight down and didn't need it. Will save that particular headache for level 2.

This solution is extremely dirty, and I should be doing proper maths to work out velocities, but it gets the lander down. I'll come back to it when I'm a bit better practiced and try for the economic achievment.